Move the Compose back to the Posts component
Once scrolling got working, I realized that the Compose message button
would only be visible from the very bottom of the list. It is not a
'sticky' footer. Now it's not part of the Messagelist, so it should
always be there.
