+++ client/zestyreactnative/src/components/message/messageList.js
@@ -6,7 +6,6 @@ import {
   View
 } from 'react-native';
 import Message from './message';
-import Button from '../common/button';
 import { getMessages } from '../../helpers/api';
 
 const ds = new ListView.DataSource({
@@ -16,34 +15,20 @@ const ds = new ListView.DataSource({
 class MessageList extends Component {
   constructor(props) {
     super(props);
-
-    this.onPressCompose = this.onPressCompose.bind(this);
-    // the footer component uses `this`, best to bind it here.
-    this.footer = this.footer.bind(this);
   }
 
   render() {
     return (
-      <View style={ [ styles.container ] }>
+      <View style={ styles.container }>
         <ListView
           style={ styles.body }
           dataSource={ ds.cloneWithRows(this.props.data) }
           renderHeader={ this.header }
           renderRow={ this.renderMessage }
           renderSeparator={ this.separator }
-          renderFooter={ this.footer }
         />
       </View>
     );
-  }
-
-  onPressCompose() {
-    var route = {
-      name: 'messageComposer',
-      callback: this.props.updateMessages
-    };
-
-    this.props.navigator.push(route);
   }
 
   renderMessage({ User, text, likes }, index) {
@@ -73,17 +58,6 @@ class MessageList extends Component {
       />
     );
   }
-
-  footer() {
-    return (
-      <View style={ styles.footer } >
-        <Button
-          style={ styles.button }
-          text={ 'Compose' }
-          onPress={ this.onPressCompose } />
-      </View>
-    );
-  }
 }
 
 const styles = StyleSheet.create({
@@ -106,14 +80,6 @@ const styles = StyleSheet.create({
     height: StyleSheet.hairlineWidth,
     backgroundColor: '#333'
   },
-  footer: {
-    flexDirection: 'row',
-    justifyContent: 'center',
-  },
-  button: {
-    width: 150,
-    marginBottom: 10
-  }
 });
 
 export default MessageList;
