+++ client/zestyreactnative/src/components/timeline/posts.js
@@ -5,6 +5,7 @@ import {
   View
 } from 'react-native';
 import MessageList from '../message/messageList';
+import Button from '../common/button';
 import { getMessages } from '../../helpers/api';
 
 var dummyMessages = [
@@ -16,6 +17,10 @@ class Posts extends Component {
   constructor(props) {
     super(props);
     this.updateMessages = this.updateMessages.bind(this);
+    this.onPressCompose = this.onPressCompose.bind(this);
+    // the footer component uses `this`, best to bind it here.
+    this.footer = this.footer.bind(this);
+
     this.state = {
       data: dummyMessages,
       error: ''
@@ -35,6 +40,26 @@ class Posts extends Component {
       .catch((err) => this.setState({ error: err }) );
   }
 
+  footer() {
+    return (
+      <View style={ styles.footer } >
+        <Button
+          style={ styles.button }
+          text={ 'Compose' }
+          onPress={ this.onPressCompose } />
+      </View>
+    );
+  }
+
+  onPressCompose() {
+    var route = {
+      name: 'messageComposer',
+      callback: this.updateMessages
+    };
+
+    this.props.navigator.push(route);
+  }
+
   render() {
     return (
       <View style={ styles.container }>
@@ -42,9 +67,9 @@ class Posts extends Component {
           <MessageList
             data={ this.state.data }
             error={ this.state.error }
-            updateMessages={ this.updateMessages }
           />
         </View>
+        { this.footer() }
       </View>
     );
   }
@@ -55,6 +80,14 @@ var styles = StyleSheet.create({
   container: {
     flex: 1,
     backgroundColor: 'white'
+  },
+  footer: {
+    flexDirection: 'row',
+    justifyContent: 'center',
+  },
+  button: {
+    width: 150,
+    marginBottom: 10
   }
 });
